GS: Fix Wsign-compare warning.

+++ pcsx2/GS/GSLocalMemory.cpp
@@ -372,7 +372,7 @@ std::vector<GSVector2i>* GSLocalMemory::GetPage2TileMap(const GIFRegTEX0& TEX0)
 	int th = std::max<int>(1 << TEX0.TH, bs.y);
 
 	// Limit the size to the maximum size of the GS memory, there's no point in mapping more than this.
-	if ((tw * th) > VM_SIZE)
+	if ((tw * th) > static_cast<int>(VM_SIZE))
 	{
 		tw = 2048;
 		th = 2048;
